refactor(share): extract AppBundleBuilder for reuse by export (#164)
